app/requirements.txt
fastai==1.0.61
fastapi==0.60.1
gunicorn==20.0.4
sentencepiece==0.1.91
scikit-learn==0.22.2
numpy==1.19.1
python-dotenv==0.14.0
gdown